+++ CHANGELOG.md
@@ -4,13 +4,14 @@ Version 1.4 (Status: DEVELOPMENT)
 * New bootstrap theme for webinterface
 * New index for webinterface (domains has now separate page - index has system status/health, like nafsdm version, kernel version and some statistics)
 * New functions to nafsdm-webinterface (upgradeCheck, settings page, logviewer)
-* New nafsdm-daemon for the master that makes sure zones managed by nafsdm exists
 * Former pythondaemon is now renamed to manager, and it's new purpose is to manage all the components of the master & upgrade the installation
 * New zone function for nafsdmctl to edit zones managed by nafsdm
 * (minor): Master upgrader restart webinterface only if it is enabled and is active
 * (minor): Move error pages to separate folder
 * (minor): Debug mode more useful in nafscli & nafsdmctl
 * (minor): Raw print function for nafsdmctl slavestatus & nafsdmctl list
+* (minor): Installation scripts now automatically detects which OS and OS version you're running
+* (not released): New nafsdm-daemon for the master that makes sure zones managed by nafsdm exists
 
 Version 1.3.1 (Status: RELEASED)
 * Bugfix: Slaves reported incorrect date to nafsdmctl slavestatus (minute instead of month under the "latest connection date" section)